Update BuildingCard.tsx
Initial damage loop would cutoff if building couldn't be destroyed with Zaps alone. This change allows the loop to continue if that isn't the case. Tested with Zap 7 and Earthquake 4 against Eagle Artillery 3. Also will not allow infinite Earthquakes to be used to calculate. (If you want this 'feature', get rid of the && q < props.spellCapacity in the if on line 55)